backend/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
sqlalchemy==2.0.44
google-api-python-client==2.193.0
google-auth-oauthlib==1.3.0
google-auth-httplib2==0.3.0
openai==2.29.0
pydantic==2.12.5
python-dotenv==1.1.1
